Move annotation parsing code to target classes
  -> Extract code to be used generically into AnnotationHelpers
  -> Create AggregationImplementation.Parser class aligned with
     its scalar protoplast.

This is first step before further refactor to show code movement before
code changes.
